Convert most closing label tests to unit tests.
See #30285
