+++ pkg/analysis_server/test/src/computer/test_all.dart
@@ -4,12 +4,14 @@
 
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
+import 'closingLabels_computer_test.dart' as closingLabels_computer_test;
 import 'import_elements_computer_test.dart' as import_elements_computer_test;
 import 'imported_elements_computer_test.dart'
     as imported_elements_computer_test;
 
 main() {
   defineReflectiveSuite(() {
+    closingLabels_computer_test.main();
     import_elements_computer_test.main();
     imported_elements_computer_test.main();
   });
